Finalized database staff!

+++ src/main/java/com/neuroval/translationApi/repository/ComparisonRepository.java
@@ -3,5 +3,5 @@ package com.neuroval.translationApi.repository;
 import com.neuroval.translationApi.model.comparison.Comparison;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComparisonRepository extends JpaRepository<Comparison, Long> {
+public interface ComparisonRepository extends JpaRepository<Comparison, Integer> {
 }
